Supp Warning

+++ scilab/routines/graphics/Axes.c
@@ -182,9 +182,9 @@ char c = (strlen(strflag) >= 3) ? strflag[2] : '1';
   if (version_flag() == 0){  
     Cscale.Waaint1[i]= pSUBWIN_FEATURE (psubwin)->axes.subint[i]+1; /*SS 01/01/03 */
     Cscale.Waaint1[0]= pSUBWIN_FEATURE (psubwin)->axes.subint[0]+1;
-    Cscale.Waaint1[1]= pSUBWIN_FEATURE (psubwin)->axes.xlim[3]; /*SS 02/01/03 */
+    Cscale.Waaint1[1]= (integer) (pSUBWIN_FEATURE (psubwin)->axes.xlim[3]); /*SS 02/01/03 */
     Cscale.Waaint1[2]= pSUBWIN_FEATURE (psubwin)->axes.subint[1]+1; 
-    Cscale.Waaint1[3]= pSUBWIN_FEATURE (psubwin)->axes.ylim[3];/*SS 02/01/03 */
+    Cscale.Waaint1[3]= (integer) (pSUBWIN_FEATURE (psubwin)->axes.ylim[3]);/*SS 02/01/03 */
 
     ticscolor=pSUBWIN_FEATURE (psubwin)->axes.ticscolor;
     textcolor=pSUBWIN_FEATURE (psubwin)->axes.textcolor;
@@ -304,9 +304,9 @@ static void aplotv1(strflag)
   if (version_flag() == 0){  
 
     Cscale.Waaint1[0]= pSUBWIN_FEATURE (psubwin)->axes.subint[0]+1;
-    Cscale.Waaint1[1]= pSUBWIN_FEATURE (psubwin)->axes.xlim[3]; /*SS 02/01/03 */
+    Cscale.Waaint1[1]= (integer) (pSUBWIN_FEATURE (psubwin)->axes.xlim[3]); /*SS 02/01/03 */
     Cscale.Waaint1[2]= pSUBWIN_FEATURE (psubwin)->axes.subint[1]+1; 
-    Cscale.Waaint1[3]= pSUBWIN_FEATURE (psubwin)->axes.ylim[3];/*SS 02/01/03 */
+    Cscale.Waaint1[3]= (integer) (pSUBWIN_FEATURE (psubwin)->axes.ylim[3]);/*SS 02/01/03 */
 
      ticscolor=pSUBWIN_FEATURE (psubwin)->axes.ticscolor;
      textcolor=pSUBWIN_FEATURE (psubwin)->axes.textcolor;
@@ -458,7 +458,7 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
     case 'u' : 
     case 'd' :
       /** Horizontal axes **/
-      barlength = Cscale.WIRect1[3]/50.0;
+      barlength =  (integer) (Cscale.WIRect1[3]/50.0);
       /** compute a format **/
       if (str == NULL && format == NULL )  
 	switch (xy_type ) {
@@ -483,7 +483,7 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
         x[0]  = ceil(Cscale.frect[0]  / (exp10( x[2]))) ; 
         x[3]=inint(x[1]-x[0]);
         while (x[3]>10)  x[3]=floor(x[3]/2);
-        Nx=x[3]+1;}
+        Nx=  (integer) (x[3]+1);}
       /**********************/
       /** loop on the ticks **/
       if (Nx==1) break; /*D.Abdemouche 16/12/2003*/
@@ -559,9 +559,9 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
 	      for ( j = 1 ; j < subtics; j++){ 
 		  vx[0] = vx[1] = XScale(vxx+dx*j);
 		  if ( pos == 'd' ) 
-		    { vy[0]= ym[0];vy[1]= ym[0] + barlength/2.0 ; }
+		    { vy[0]= ym[0];vy[1]=  (integer) (ym[0] + barlength/2.0) ; }
 		  else 
-		    { vy[0]= ym[0];vy[1]= ym[0] - barlength/2.0; }
+		    { vy[0]= ym[0];vy[1]=  (integer) (ym[0] - barlength/2.0) ; }
 		  C2F(dr)("xsegs","v", vx, vy, &ns,&style,&iflag,PI0,PD0,PD0,PD0,PD0,0L,0L);
 		}
 	    } 
@@ -577,9 +577,9 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
                  if ( vxx-dx*j > xmin){
 		  vx[0] = vx[1] = XScale(vxx-dx*j);
 		  if ( pos == 'd' ) 
-		    { vy[0]= ym[0];vy[1]= ym[0] + barlength/2.0 ; }
+		    { vy[0]= ym[0];vy[1]=  (integer) (ym[0] + barlength/2.0) ; }
 		  else 
-		    { vy[0]= ym[0];vy[1]= ym[0] - barlength/2.0; }
+		    { vy[0]= ym[0];vy[1]=  (integer) (ym[0] - barlength/2.0) ; }
 		  C2F(dr)("xsegs","v", vx, vy, &ns,&style,&iflag,PI0,PD0,PD0,PD0,PD0,0L,0L);
 		 }}
 	    } 
@@ -593,9 +593,9 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
                  if ( vxx+dx*j < xmax){
 		  vx[0] = vx[1] = XScale(vxx+dx*j);
 		  if ( pos == 'd' ) 
-		    { vy[0]= ym[0];vy[1]= ym[0] + barlength/2.0 ; }
+		    { vy[0]= ym[0];vy[1]=  (integer) (ym[0] + barlength/2.0) ; }
 		  else 
-		    { vy[0]= ym[0];vy[1]= ym[0] - barlength/2.0; }
+		    { vy[0]= ym[0];vy[1]=  (integer) (ym[0] - barlength/2.0) ; }
 		  C2F(dr)("xsegs","v", vx, vy, &ns,&style,&iflag,PI0,PD0,PD0,PD0,PD0,0L,0L);
 		 }}
 	    } 
@@ -607,7 +607,7 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
     case 'r' : 
     case 'l' :
       /** Vertical axes **/
-      barlength = Cscale.WIRect1[2]/75.0;
+      barlength =  (integer) (Cscale.WIRect1[2]/75.0);
       if (str == NULL &&  format == NULL )  
 	switch (xy_type ) {
 	case 'v' : ChoixFormatE1(c_format,y,Ny);break;
@@ -631,7 +631,7 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
         y[0]  = ceil(Cscale.frect[1]  / (exp10( y[2]))) ; 
         y[3]=inint(y[1]-y[0]);
         while (y[3]>10)  y[3]=floor(y[3]/2);
-        Ny=y[3]+1;}
+        Ny= (integer) (y[3]+1);}
       /**********************/
       /** loop on the ticks **/
        if (Ny==1) break; /*D.Abdemouche 16/12/2003*/
@@ -706,9 +706,9 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
 		{
 		  vy[0] = vy[1] = YScale(vxx+dy*j);
 		  if ( pos == 'r' ) 
-		    { vx[0]= xm[0];vx[1]= xm[0] + barlength/2.0 ; }
+		    { vx[0]= xm[0];vx[1]= (integer) (xm[0] + barlength/2.0) ; }
 		  else 
-		    { vx[0]= xm[0];vx[1]= xm[0] - barlength/2.0; }
+		    { vx[0]= xm[0];vx[1]= (integer) (xm[0] - barlength/2.0) ; }
 		  C2F(dr)("xsegs","v", vx, vy, &ns,&style,&iflag,PI0,PD0,PD0,PD0,PD0,0L,0L);
 		}
 	    }
@@ -724,9 +724,9 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
                  if ( vxx-dy*j > ymin){
 		  vy[0] = vy[1] = YScale(vxx-dy*j);
 		  if ( pos == 'r' ) 
-		    { vx[0]= xm[0];vx[1]= xm[0] + barlength/2.0 ; }
+		    { vx[0]= xm[0];vx[1]= (integer) (xm[0] + barlength/2.0) ; }
 		  else 
-		    { vx[0]= xm[0];vx[1]= xm[0] - barlength/2.0; }
+		    { vx[0]= xm[0];vx[1]= (integer) (xm[0] - barlength/2.0) ; }
 		  C2F(dr)("xsegs","v", vx, vy, &ns,&style,&iflag,PI0,PD0,PD0,PD0,PD0,0L,0L);
 		 }}
 	    }
@@ -740,9 +740,9 @@ void Sci_Axis(pos,xy_type,x,nx,y,ny,str,subtics,format,fontsize,textcolor,ticsco
                  if ( vxx+dy*j < ymax){
 		  vy[0] = vy[1] = YScale(vxx+dy*j);
 		  if ( pos == 'r' ) 
-		    { vx[0]= xm[0];vx[1]= xm[0] + barlength/2.0 ; }
+		    { vx[0]= xm[0];vx[1]= (integer) (xm[0] + barlength/2.0) ; }
 		  else 
-		    { vx[0]= xm[0];vx[1]= xm[0] - barlength/2.0; }
+		    { vx[0]= xm[0];vx[1]= (integer) (xm[0] - barlength/2.0) ; }
 		  C2F(dr)("xsegs","v", vx, vy, &ns,&style,&iflag,PI0,PD0,PD0,PD0,PD0,0L,0L);
 		 }}
 	    }
